+++ app/build.gradle.kts
@@ -9,18 +9,28 @@ android {
   namespace = "com.nexters.bandalart.android"
 
   buildFeatures {
+    compose = true
     buildConfig = true
+  }
+  composeOptions {
+    kotlinCompilerExtensionVersion = libs.versions.androidx.compose.compiler.get()
   }
 }
 
 dependencies {
   coreLibraryDesugaring(libs.desugar.jdk)
   implementations(
+    projects.core.data,
+    projects.core.domain,
+    projects.core.designsystem,
     libs.androidx.splash,
     libs.androidx.startup,
+    libs.androidx.appcompat,
+    libs.androidx.core,
+    libs.androidx.hilt.compose.navigation,
+    libs.androidx.splash,
     libs.timber,
-    projects.data,
-    projects.domain,
-    projects.presentation,
+    libs.bundles.androidx.compose,
+    libs.bundles.androidx.lifecycle,
   )
 }
