+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,18 @@
 
     </provider>
 
+    <activity
+      android:name="com.nexters.bandalart.android.MainActivity"
+      android:exported="true"
+      android:theme="@style/Theme.Bandalart.Splash">
+
+      <intent-filter>
+        <action android:name="android.intent.action.MAIN" />
+        <category android:name="android.intent.category.LAUNCHER" />
+      </intent-filter>
+
+    </activity>
+
   </application>
 
 </manifest>
