+++ app/src/main/kotlin/com/nexters/bandalart/android/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.nexters.bandalart.android
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.material3.Text
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.nexters.bandalart.android.ui.theme.BandalartTheme
+
+class MainActivity : ComponentActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    installSplashScreen()
+    setContent {
+      BandalartTheme {
+        Text(text = "Bandalart Hello")
+      }
+    }
+  }
+}
