+++ app/src/main/res/values/splash.xml
