+++ build.gradle.kts
@@ -11,6 +11,7 @@ plugins {
   alias(libs.plugins.android.application) apply false
   alias(libs.plugins.android.library) apply false
   alias(libs.plugins.android.hilt) apply false
+  alias(libs.plugins.kotlin.android) apply false
 }
 
 buildscript {
