+++ core/designsystem/src/main/kotlin/com/nexters/bandalart/android/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.presentation.ui.theme
+package com.nexters.bandalart.android.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
