--- presentation/src/main/AndroidManifest.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.nexters.bandalart.android.presentation">
-
-  <application android:theme="@style/Theme.Bandalart">
-    <activity
-      android:name="com.nexters.bandalart.android.presentation.MainActivity"
-      android:exported="true"
-      android:theme="@style/Theme.Bandalart.Splash">
-
-      <intent-filter>
-        <action android:name="android.intent.action.MAIN" />
-        <category android:name="android.intent.category.LAUNCHER" />
-      </intent-filter>
-
-    </activity>
-
-  </application>
-
-</manifest>
